1.08b du 15 juin 2013 (et 16 juin 2013)
Hi Robin, I re-upload 1.08b as an embarrassing 2^31=2147483647
(hard for a power of two to be odd, don't you think?) prominently
figured on page 4 of the pdf documentation from yesterday's
upload. I don't know if this will provide you with some relief,
but I have decided to stop here (at least for the summer) further
development of the package. I have re-entered the last
announcement text in the upload form. Thanks a lot, and best
wishes. Jean-Francois.

   (1.08b)
   1. a bug with spaces inside \xinttheexpr...\relax is fixed.
   2. some new improvements in the handling of floating point
      numbers.
   3. the previously non-documented possibilities for using
      count registers and infix operations with them directly
      in the arguments to the macros of xintfrac.sty are
      explained.