1.09f du 4 novembre 2013
New features:

 * \xintZapSpaces /completely expandably/ removes both leading and
   ending space tokens (char code 32) from its input. The input is
   not altered in any other way. The variant \xintZapSpacesB
   additionally transforms <spaces>{braced}<spaces in braced (now
   unbraced...). \xintZapFirstSpaces and \xintZapLastSpaces do
   what their name indicate. The input may contain unmatched \if,
   \fi tokens etc... but top level space tokens are assumed to be
   of character code 32, as this works with delimited macros.

 * the completely expandable \xintCSVtoList now strips out all
   spaces around commas or at the start and end of the list of
   the comma separated items (using \xintZapSpacesB).

 * the non completely expandable \xintFor et al. all similarly now
   removes all spaces around commas.

 * \xintFor et al. are extended to accept all macro parameters
   from #1 to #9 (in arbitrary order); this will prove very useful
   for those who need to typeset a tabular on 9-dimensional paper.

I plan on issuing at some point a spin-off package with the macros
above (all those described in the documentation section on the
"Utilities" provided by xint.sty). Perhaps \xintZapSpaces could be
made into a little package of its own; or the package should
contain only what is necessary for \xintFor and its variants.
Feedback will be welcome on whether this is worthwile and if the
thing should, as the xint bundle packages, be made available to
all macro formats or only as a LaTeX package.